Added lsignals
Added lsignals header file to eventually use the signal slot design pattern in order to update OutputComm.cpp, which, will be used to package together images and slippi data together. yaml-cpp was removed as nlohmann/json was found to be used in Slippi, no need to have two json parsing sources to be needed.