ADDED FIRST STAGE OUTPUT
Originally the OutputComm was setup to act as a server and handle clients, correlate messages, etc, but not anymore. The purpose now is to allow two instances spawned, such as the 4 instances spawned for InputComm and output the messages to anyone listening on the designated IP + port. This in turn allows the other side to handle the messages while Dolphin doesn't need anymore additional overhead correlating messages together.