Added Slippi network data
Slippi data was only being sent over the network during live sessions, the latest update adds network communication during replays. The end message is unable to be sent right now however it'll be added in the future. It's not that high of a priority as it'll only indicate LRA exit which can be flagged before extraction.